refactor(IPVC-2276): pull out open_file into module
